Resolved issue with the converted value in the database

+++ app/Casts/MoneyCast.php
@@ -26,6 +26,6 @@ class MoneyCast implements CastsAttributes
     public function set(Model $model, string $key, mixed $value, array $attributes): mixed
     {
         // Transform the float into an integer for storage
-        return round(floatval($value) * 100);
+        return (int) round(floatval($value) * 100);
     }
 }
